avoid chunked transfers in putBlob when possible
This is to avoid Downloads.jl from initiating chunked transfer when a file IOStream is supplied to `putBlob`.

In 0.4.0, we have switched to using Downloads.jl as the HTTP client instead of HTTP.jl. This issue happens because Downloads.jl
is not able to determine the content length when an IO instance is provided to upload data from, and switches to using a chunked transfer encoding. And the putBlob Azure API endpoint does not support chunked transfers.

The solution is to not use chunked transfers for putBlob. This PR (https://github.com/JuliaLang/Downloads.jl/pull/167) to Downloads.jl would let it also look into the content-length header that we supply to determine the data size to expect in the IO instance we pass.

But until the Downloads.jl PR is merged (and backported), we need this change in Azure.jl to detect IO handles that point to locally mappable files and supply the correct content length and `IOBuffer` of memory mapped file contents as data to upload instead.

fixes: #25